rename RemoveCombinationC command and handler to DeleteCombination
