+++ src/PrestaShopBundle/Resources/config/services/adapter/product/_combination.yml
@@ -168,13 +168,13 @@ services:
             - name: tactician.handler
               command: PrestaShop\PrestaShop\Core\Domain\Product\Combination\Command\RemoveAllCombinationImagesCommand
 
-    prestashop.adapter.product.combination.command_handler.remove_combination_handler:
-        class: PrestaShop\PrestaShop\Adapter\Product\Combination\CommandHandler\RemoveCombinationCommandHandler
+    prestashop.adapter.product.combination.command_handler.delete_combination_handler:
+        class: PrestaShop\PrestaShop\Adapter\Product\Combination\CommandHandler\DeleteCombinationHandler
         arguments:
             - '@prestashop.adapter.product.combination.update.combination_deleter'
         tags:
             - name: tactician.handler
-              command: PrestaShop\PrestaShop\Core\Domain\Product\Combination\Command\RemoveCombinationCommand
+              command: PrestaShop\PrestaShop\Core\Domain\Product\Combination\Command\DeleteCombinationCommand
 
     prestashop.adapter.product.combination.command_handler.bulk_delete_combination_handler:
         class: PrestaShop\PrestaShop\Adapter\Product\Combination\CommandHandler\BulkDeleteCombinationHandler
